+++ src/LASRcore/LAScatalog.cpp
@@ -353,8 +353,8 @@ bool LAScatalog::write_vpc(const std::string& vpcfile, const CRS& crs, bool abso
     output << "      \"pc:type\": " << "\"lidar\"" << ","<< std::endl;
     if (index) output << "      \"index:indexed\": " << "true," << std::endl;
     output << "      \"proj:bbox\": [" << std::fixed << std::setprecision(3) << bbox.minx << ", " << bbox.miny << ", " << bbox.maxx << ", " << bbox.maxy << "],"<< std::endl;
-    if (!wkt.empty()) output << "      \"proj:wtk2\": " << wkt << "," << std::endl;
-    if (epsg != 0) output << "      \"proj:epsg\": " << epsg << std::endl;
+    if (!wkt.empty()) output << "      \"proj:wkt2\": " << wkt << "," << std::endl;
+    if (epsg != 0) output << "      \"proj:epsg\": " << epsg << "," << std::endl;
     output << "    }," << std::endl;
     output << "    \"links\": []," << std::endl;
     output << "    \"assets\": {" << std::endl;