+++ NEWS.md
@@ -6,6 +6,13 @@
 - New: Added new stage `write_lax`. This stage was automatically added by the engine but can now be explicitly added by users.
 - Change: The package no longer assigns `set_parallel_strategy(concurrent_points(half_core()))` when loading. Instead, if nothing is provided, this is interpreted as `concurrent_points(half_core())`. Thus, users can now write `exec(pipeline, on = file, ncores = 8)`. The engine will now respect `ncores = 8` because no global settings were assigned. The multi-threading vignette has been updated.
 
+# lasR 0.5.6
+
+- Fix: `reader_las()` with COPC files, depth query (`-max_depth`), and buffer. The depth query was not performed at all. The fix is temporary: it breaks the progress bar of `reader_las()` but this is a less serious bug.
+- Fix: `reader_las()` with very large files.
+- Fix: `load_raster()` is thread-safe
+- New: `rasterize()` accepts a new argument `default_value`.  When rasterizing with an operator and a filter (e.g. `-keep_z_above 2`) some pixels that are covered by points may no longer contain any point that pass the filter criteria and are assigned NA. To differentiate NAs from non covered pixels and NAs from covered pixels but without point that pass the filter, the later case can be assigned another value such as 0.
+
 # lasR 0.5.5
 
 - Fix: #50 `write_vpc()` properly reprojects the bounding boxes in WGS84